componentize browser utility functions
