+++ src/utils/is-mobile.js
@@ -1,0 +1,3 @@
+export default function isMobile (agent = navigator.userAgent) {
+	return /Android|iPhone|iPad|iPod/i.test(agent)
+}
