+++ src/utils/transform-supported.js
@@ -1,0 +1,4 @@
+export default function transformSupported () {
+	const style = document.documentElement.style
+	return 'transform' in style || 'WebkitTransform' in style
+}
