+++ src/utils/transition-supported.js
@@ -1,0 +1,4 @@
+export default function transitionSupported () {
+	const style = document.documentElement.style
+	return 'transition' in style || 'WebkitTransition' in style
+}
