+++ test/utils/transform-supported.spec.js
@@ -1,0 +1,9 @@
+import transformSupported from '../../src/utils/transform-supported'
+
+describe('Utilities', () => {
+	describe('transformSupported()', () => {
+		it('should return true', () => {
+			expect(transformSupported()).to.be.true
+		})
+	})
+})
