+++ test/utils/transition-supported.spec.js
@@ -1,0 +1,9 @@
+import transitionSupported from '../../src/utils/transition-supported'
+
+describe('Utilities', () => {
+	describe('transitionSupported()', () => {
+		it('should return true', () => {
+			expect(transitionSupported()).to.be.true
+		})
+	})
+})
